rand done and fixed some bug

+++ widget.cpp
@@ -29,9 +29,24 @@ Widget::~Widget()
 void Widget::mousePressEvent(QMouseEvent *event){
     //draw=true;
     if (ui->addVertexBtn->isChecked()){
-    Vertex* v=new Vertex(event->x(), event->y(), ui->label);
+        addVertexAction(event->x(),event->y());
+//    Vertex* v=new Vertex(event->x(), event->y(), ui->label);
+//    //qDebug()<<event->x()<<event->y();
+//    m.addVertex(v);
+//    connect(v,SIGNAL(deleteAction(Vertex*, bool &)), this, SLOT(rightClickVertex(Vertex*, bool &)));
+//    connect(ui->addEdgeBtn, SIGNAL(toggled(bool)),v, SLOT(catchEdgeBtnState(bool)));
+//    connect(v,SIGNAL(addEdgeAction(Vertex *, const bool &)),this,SLOT(leftClickVertex(Vertex *, const bool &)));
+//    connect(ui->startCalcalationBtn,SIGNAL(toggled(bool)),v,SLOT(catchStartBtnState(bool)));
+//    connect(v,SIGNAL(startCalAction(Vertex*)),this,SLOT(startCalcalationAction(Vertex*)));
+//    connect(v,SIGNAL(showP(Vertex*)),this,SLOT(showPathAction(Vertex*)));
+//    connect(ui->showPathBtn,SIGNAL(toggled(bool)),v,SLOT(catchShowPathBtnState(bool)));
+    }
+}
+
+void Widget::addVertexAction(const int & x, const int & y)
+{
+    Vertex* v=new Vertex(x, y, ui->label);
     m.addVertex(v);
-    //v->showTentativeDistance();
     connect(v,SIGNAL(deleteAction(Vertex*, bool &)), this, SLOT(rightClickVertex(Vertex*, bool &)));
     connect(ui->addEdgeBtn, SIGNAL(toggled(bool)),v, SLOT(catchEdgeBtnState(bool)));
     connect(v,SIGNAL(addEdgeAction(Vertex *, const bool &)),this,SLOT(leftClickVertex(Vertex *, const bool &)));
@@ -39,10 +54,7 @@ void Widget::mousePressEvent(QMouseEvent *event){
     connect(v,SIGNAL(startCalAction(Vertex*)),this,SLOT(startCalcalationAction(Vertex*)));
     connect(v,SIGNAL(showP(Vertex*)),this,SLOT(showPathAction(Vertex*)));
     connect(ui->showPathBtn,SIGNAL(toggled(bool)),v,SLOT(catchShowPathBtnState(bool)));
-    //connect(v,SIGNAL(moveEdgeAction(Vertex *)),this,SLOT(moveVertex(Vertex *)));
-    }
 }
-
 void Widget::on_addVertexBtn_toggled(bool checked)
 {
     if (checked)
@@ -60,9 +72,18 @@ void Widget::leftClickVertex(Vertex* v, const bool & state){
     if (state){
         edgePoint.push_back(v);
         if (edgePoint.size()>=2){
+            if (m.edgeExist(edgePoint.front(),edgePoint.back())){
+                QMessageBox msg;
+                msg.setText("Edge linked two selected vertex is already existed");
+                msg.exec();
+                edgePoint.back()->unSelect();
+                edgePoint.pop_back();
+                return;
+            }
             bool ok=false;
             int distance=QInputDialog::getInt(this, "", "Plaese input the distance",0,0,INT_MAX,1,&ok);
             if (ok){
+
                 Edge * e=new Edge(edgePoint.front(),edgePoint.back(), distance,ui->label);
                 e->lower();
                 connect(e,SIGNAL(deleteAction(Edge*,bool&)),this,SLOT(deleteEdgeAction(Edge*,bool&)));
@@ -128,4 +149,53 @@ void Widget::on_clearBtn_clicked()
     ui->addVertexBtn->setEnabled(true);
     ui->addEdgeBtn->setEnabled(true);
     m.clearResult();
+}
+
+void Widget::on_stepBtn_clicked()
+{
+
+}
+
+void Widget::on_randGenBtn_clicked()
+{
+    bool vertexok=false;
+    int numOfVertex=QInputDialog::getInt(this, "", "Plaese input the number of vertex:",1,1,INT_MAX,1,&vertexok);
+    if (vertexok){
+        bool edgeok=false;
+        double denOfEdge=QInputDialog::getDouble(this,"","Please input the density of edges [0-1]:",0,0,1,2,&edgeok);
+        if (edgeok){
+                bool maxWeightok=false;
+                int maxWeight=QInputDialog::getInt(this, "", "Plaese input the maximum edge weight",0,0,INT_MAX,1,&maxWeightok);
+                if (maxWeightok){
+                    srand (time(NULL));
+                    for (int v=0;v<numOfVertex;++v){
+                        int xpos=rand()%720+40;
+                        int ypos=rand()%460+40;
+                        addVertexAction(xpos,ypos);
+                    }
+                    //qDebug()<<"pass add vertex";
+                    int numOfEdge=numOfVertex*(numOfVertex-1)/2*denOfEdge;
+                    for (int e=0;e<numOfEdge;++e){
+                        //qDebug()<<"before get";
+                        Vertex* firstVertex=m.getVertex(rand()%numOfVertex);
+                        Vertex* secondVertex=m.getVertex(rand()%numOfVertex);
+                        //qDebug()<<"after get";
+                        if (!m.edgeExist(firstVertex,secondVertex)){
+                            Edge * e=new Edge(firstVertex,secondVertex, rand()%(maxWeight+1),ui->label);
+                            e->lower();
+                            connect(e,SIGNAL(deleteAction(Edge*,bool&)),this,SLOT(deleteEdgeAction(Edge*,bool&)));
+                            m.addEdge(e);
+                            std::list<Vertex*> edgeTwoPoint;
+                            edgeTwoPoint.push_back(firstVertex);
+                            edgeTwoPoint.push_back(secondVertex);
+                            m.connectVertex(edgeTwoPoint);
+                        }else{
+                            --e;
+                        }
+                    }
+                    //qDebug()<<"pass add edge";
+                }
+        }
+
+    }
 }
